telegram_bot/requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
apscheduler-di==0.1.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
rodi==2.0.6
six==1.16.0
typing_extensions==4.12.2
tzlocal==5.2
yarl==1.9.4
